+++ Changelog.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Remove unused console.log. Completed on 2025-10-06
+
 ### Fixed
 
 ### Security