+++ Changelog.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Bug fix webbunjs engine insert empty value into the sessiion table issue. Completed on 2025-10-06
+- Fixed some typos in dependency module versions. Completed on 2025-10-06
 
 ### Security