+++ Changelog.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.7] - 2025-10-07
+
+### Summary
+
+- Focus on developing the deskelectronjs engine to make it more powerful.
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+- Fix write `route.json` file permission issue cause entire service corrupted in app.js. Completed on 2025-10-07
+
+### Security
+
+[1.0.7]: https://github.com/wkloh76/oricommjs_v2/releases/tag/1.0.7
+
 ## [1.0.6] - 2025-10-05
 
 ### Summary